requirements.txt
python-dotenv==1.1.1
chromadb==1.0.15
openai==1.97.1
tiktoken==0.9.0
fastapi==0.116.1
uvicorn==0.35.0 